requirements.txt
langchain
langchain-community
langchain-openai
langchain-text-splitters
langchain-experimental
langgraph
langchain-mcp-adapters
openai

faiss-cpu
pymupdf

gradio
python-dotenv
pandas
tqdm
networkx
wikipedia
duckduckgo-search

langchain-graph-retriever
graph-retriever
graph-rag-example-helpers

# B2B Account Intelligence Agent
anthropic
tavily-python
